requirements.txt
apache-airflow-providers-amazon
soda-core-redshift
pendulum
pandas
boto3